requirements.txt
pywebview>=5.0
psutil>=5.9
Pillow>=10.0.0
